refactor(lex): Split out input cursor
