+++ src/parse/mod.rs
@@ -1,13 +1,15 @@
-pub mod features;
-
 mod arg_matcher;
-pub mod matches;
+mod lexer;
 mod parser;
 mod validator;
 
+pub mod features;
+pub mod matches;
+
 pub(crate) use self::arg_matcher::ArgMatcher;
+pub(crate) use self::lexer::Input;
 pub(crate) use self::matches::{MatchedArg, SubCommand};
-pub(crate) use self::parser::{Input, ParseState, Parser};
+pub(crate) use self::parser::{ParseState, Parser};
 pub(crate) use self::validator::Validator;
 
 pub use self::matches::{ArgMatches, Indices, OsValues, ValueSource, Values};
